:sparkles: Enable to run `pin()` with an existing socket
